refactor(chat): split voice-input + scroll behaviors out of ChatTabView
ChatTabView was the second-largest file in the repo (1042 lines) and
mixed three responsibilities: the chat composer + message list view,
the voice-input mic plumbing (heartbeat task, event consumer, toggle),
and a couple of pure-static behavior calculators.

This PR pulls the obvious-cohort code into siblings:

- `ChatTabView+VoiceInput.swift` (197 lines): the
  `SpeechPartialTranscriptBuffer` actor-equivalent, the
  `mergedSpeechInput` / `speechFailureInput` static helpers, the
  heartbeat / event-consumer / toggleRecording lifecycle. SwiftUI
  requires `@State` to live on the containing struct, so the field
  declarations stay on `ChatTabView` (demoted from `private` to
  internal so the extension can read them); only the lifecycle
  methods move.
- `ChatBehaviors.swift` (32 lines): `ChatScrollBehavior` and
  `SessionListEdgeSwipeBehavior` — pure-static threshold logic that
  has nothing to do with `ChatTabView`'s view body.

ChatTabView.swift: 1042 → 844 lines (-19%).

What stays:
- `MessageGroupItem` (file-private, used only inside the view body)
- `BottomMarkerMinYPreferenceKey` (same)
- `TurnActivityRowView` (depends on a fileprivate nested type
  `ChatTabView.TurnActivity` — splitting would require leaking the
  type to module scope)

Tests: TEST SUCCEEDED.
